tests/robot/requirements.txt
# pip install -r requirements.txt
pyserial==3.4

